moved analysis package to lens.model
