+++ autolens/interferometer/model/aggregator.py
@@ -7,9 +7,9 @@ import autogalaxy as ag
 from autogalaxy.analysis.aggregator.aggregator import _interferometer_from
 
 from autolens.interferometer.fit_interferometer import FitInterferometer
-from autolens.analysis.aggregator.aggregator import AbstractAgg
+from autolens.lens.model.aggregator.aggregator import AbstractAgg
 
-from autolens.analysis.aggregator.aggregator import _tracer_from
+from autolens.lens.model.aggregator.aggregator import _tracer_from
 
 
 def _fit_interferometer_from(
