+++ autolens/interferometer/model/visualizer.py
@@ -1,7 +1,7 @@
 from autolens.plot.fit_interferometer_plotters import FitInterferometerPlotter
-from autolens.analysis.visualizer import Visualizer
+from autolens.lens.model.visualizer import Visualizer
 
-from autolens.analysis.visualizer import plot_setting
+from autolens.lens.model.visualizer import plot_setting
 
 
 class VisualizerInterferometer(Visualizer):
