+++ autolens/lens/model/aggregator/__init__.py
@@ -4,13 +4,13 @@ from autogalaxy.analysis.aggregator.aggregator import ImagingAgg
 from autogalaxy.analysis.aggregator.aggregator import _interferometer_from
 from autogalaxy.analysis.aggregator.aggregator import InterferometerAgg
 
-from autolens.analysis.aggregator.aggregator import _tracer_from
-from autolens.analysis.aggregator.aggregator import TracerAgg
+from autolens.lens.model.aggregator.aggregator import _tracer_from
+from autolens.lens.model.aggregator.aggregator import TracerAgg
 
 from autolens.imaging.model.aggregator import _fit_imaging_from
 from autolens.imaging.model.aggregator import FitImagingAgg
 
-from autolens.analysis.aggregator.aggregator import _interferometer_from
-from autolens.analysis.aggregator.aggregator import FitInterferometerAgg
+from autolens.lens.model.aggregator.aggregator import _interferometer_from
+from autolens.lens.model.aggregator.aggregator import FitInterferometerAgg
 
-from autolens.analysis.aggregator.aggregator import SubhaloAgg
+from autolens.lens.model.aggregator.aggregator import SubhaloAgg
