+++ test_autolens/lens/model/test_visualizer.py
@@ -5,7 +5,7 @@ from os import path
 import pytest
 from autoconf import conf
 import autolens as al
-from autolens.analysis import visualizer as vis
+from autolens.lens.model import visualizer as vis
 
 directory = path.dirname(path.realpath(__file__))
